requirements.txt
azure-devops==7.1.0b4
msrest>=0.7.1,<0.8.0

